Refactor auth package
