--- auth/src/main/java/org/aerogear/android/ags/auth/AuthServiceConfig.java
@@ -1,7 +1,0 @@
-package org.aerogear.android.ags.auth;
-
-/**
- * Authentication service configuration
- */
-public class AuthServiceConfig {
-}
