--- auth/src/main/java/org/aerogear/android/ags/auth/RoleType.java
@@ -1,8 +1,0 @@
-package org.aerogear.android.ags.auth;
-
-/**
- * Keycloak role types
- */
-public enum RoleType {
-    REALM, CLIENT
-}
