--- auth/src/main/java/org/aerogear/android/ags/auth/credentials/ICredential.java
@@ -1,7 +1,0 @@
-package org.aerogear.android.ags.auth.credentials;
-
-/**
- * Base interface for credential objects.
- */
-public interface ICredential {
-}
