+++ auth/src/main/java/org/aerogear/mobile/auth/AuthServiceConfig.java
@@ -1,0 +1,7 @@
+package org.aerogear.mobile.auth;
+
+/**
+ * Authentication service configuration
+ */
+public class AuthServiceConfig {
+}
