+++ auth/src/main/java/org/aerogear/mobile/auth/AuthenticationException.java
@@ -1,0 +1,19 @@
+package org.aerogear.mobile.auth;
+
+/**
+ * Exception thrown when an error occurs authenticating a user.
+ */
+public class AuthenticationException extends Exception {
+
+    public AuthenticationException(final Throwable cause) {
+        super(cause);
+    }
+
+    public AuthenticationException(final String message) {
+        super(message);
+    }
+
+    public AuthenticationException(final String message, final Throwable cause) {
+        super(message, cause);
+    }
+}
