+++ auth/src/main/java/org/aerogear/mobile/auth/RoleType.java
@@ -1,0 +1,8 @@
+package org.aerogear.mobile.auth;
+
+/**
+ * Keycloak role types
+ */
+public enum RoleType {
+    REALM, CLIENT
+}
