+++ auth/src/main/java/org/aerogear/mobile/auth/credentials/ICredential.java
@@ -1,0 +1,7 @@
+package org.aerogear.mobile.auth.credentials;
+
+/**
+ * Base interface for credential objects.
+ */
+public interface ICredential {
+}
